Batch 1 for javadoc.

+++ src/main/java/com/zeitheron/hammercore/client/model/simple/ModelOpcodes.java
@@ -1,14 +1,17 @@
 package com.zeitheron.hammercore.client.model.simple;
 
+/**
+ * Some opcodes for simple models.
+ */
 public interface ModelOpcodes
 {
-	public static final int DFACE = 0x00;
-	public static final int EFACE = 0x01;
-	public static final int COLOR = 0x02;
-	public static final int BOUNDS = 0x03;
-	public static final int DRAW = 0x04;
-	public static final int DFACES = 0x05;
-	public static final int EFACES = 0x06;
-	public static final int ITEX = 0x07;
-	public static final int INAME = 0x08;
+	public static final byte DFACE = 0x00;
+	public static final byte EFACE = 0x01;
+	public static final byte COLOR = 0x02;
+	public static final byte BOUNDS = 0x03;
+	public static final byte DRAW = 0x04;
+	public static final byte DFACES = 0x05;
+	public static final byte EFACES = 0x06;
+	public static final byte ITEX = 0x07;
+	public static final byte INAME = 0x08;
 }
